completed the scraping lambdas. moved some files around and deleted old files
