+++ AWS_codebase/StockScraper/collect5stocks.py
@@ -1,43 +1,13 @@
-import json
 import boto3
-import traceback
 from external_api_comm import get_stockprice, convert_dict_to_df
 
-getMethod = 'GET'
-postMethod = 'POST'
-healthPath = '/health'
-scrapePath = '/scrape/prices'
 
 dynamodb = boto3.resource('dynamodb', region_name='us-east-2')
 table = dynamodb.Table('StockPrices')
 
 
 def lambda_handler(event, context):
-    try:
-        httpMethod = event['httpMethod']
-        path = event['path']
-        # first scenario, health check
-        if httpMethod == getMethod and path == healthPath:
-            response = respond(None, 'successful')
-        # second scenario, scrape stocks
-        elif httpMethod == postMethod and path == scrapePath:
-            failed_stocks = write_stocks(json.loads(event['body']))
-            response = respond(None, 'successful') if not failed_stocks else respond(ValueError(f'Failed to scrape {failed_stocks}')) 
-        else:
-            response = respond(ValueError(f'Unsupported method/path: {httpMethod}/{path}'))
-        return response
-    except Exception as e:
-        return respond(traceback.format_exc())
-    
-
-def respond(err, res=None):
-    return {
-        'statusCode': '400' if err else '200',
-        'body': str(err) if err else json.dumps(res),
-        'headers': {
-            'Content-Type': 'application/json',
-        },
-    }
+    write_stocks(event['body'])
 
 
 def write_stocks(stock_list):
@@ -66,4 +36,3 @@ def write_stocks(stock_list):
                 # write the data to dynamodb
                 batch.put_item(Item=row_dict)
     return failed_stocks
-
